# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "description": "모든 HTTP 오류가 공유하는 응답 형식.",
        "properties": {
          "detail": {
            "description": "사용자와 개발자가 읽을 수 있는 한글 오류 설명",
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "description": "클라이언트가 분기 처리할 수 있는 영문 오류 코드",
            "title": "Error Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "요청 필드 검증 오류가 있을 때만 포함하는 세부 정보",
            "title": "Errors"
          },
          "status": {
            "const": "ERROR",
            "default": "ERROR",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "ConfirmRequest": {
        "description": "사용자가 확인 화면에서 검토·수정한 뒤 보내는 확정 요청.\n\nAI 서비스는 상태를 보관하지 않습니다. 백엔드가 ``/from-image`` 응답을 들고 있다가\n사용자 수정본과 함께 그대로 되돌려주면 됩니다.",
        "properties": {
          "approved": {
            "default": true,
            "description": "false면 아무것도 등록하지 않습니다.",
            "title": "Approved",
            "type": "boolean"
          },
          "calendar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CalendarDraft"
              },
              {
                "type": "null"
              }
            ],
            "description": "사용자가 수정한 일정 초안. 생략하면 gift_data 로 다시 계산합니다."
          },
          "calendar_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Id"
          },
          "gift_data": {
            "$ref": "#/components/schemas/GiftData",
            "description": "사용자가 수정한 기록. records 에 남은 항목만 저장됩니다."
          },
          "google_access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "사용자 Google OAuth access token. 없으면 서버 설정값을 씁니다.",
            "title": "Google Access Token"
          },
          "register_calendar": {
            "default": true,
            "title": "Register Calendar",
            "type": "boolean"
          },
          "workflow_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "gift_data"
        ],
        "title": "ConfirmRequest",
        "type": "object"
      },
      "ConfirmResponse": {
        "description": "확정 처리 결과.",
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "calendar_info": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "gift_data": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "noti_info": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "approved",
          "gift_data",
          "calendar_info",
          "noti_info"
        ],
        "title": "ConfirmResponse",
        "type": "object"
      },
      "Gender": {
        "description": "상대방 성별. 추천 카테고리에 영향을 줍니다.",
        "enum": [
          "male",
          "female",
          "unknown"
        ],
        "title": "Gender",
        "type": "string"
      },
      "GiftAgentResponse": {
        "description": "네 비동기 작업 결과를 합친 최종 HTTP 응답.",
        "properties": {
          "calendar_info": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "gift_data": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "noti_info": {
            "$ref": "#/components/schemas/PreparedData"
          },
          "recommend_gift_info": {
            "$ref": "#/components/schemas/GiftRecommendationInfo"
          },
          "requires_confirmation": {
            "default": false,
            "description": "true면 아직 캘린더에 등록되지 않았습니다. 사용자 확인 후 /confirm 을 호출하세요.",
            "title": "Requires Confirmation",
            "type": "boolean"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "네 작업 결과를 연결하는 ID. /confirm 에 그대로 돌려주세요.",
            "title": "Workflow Id"
          }
        },
        "required": [
          "gift_data",
          "calendar_info",
          "noti_info",
          "recommend_gift_info"
        ],
        "title": "GiftAgentResponse",
        "type": "object"
      },
...
      "GiftDataRequest": {
        "description": "선물데이터 직접 전달 API의 요청 본문.",
        "example": {
          "gift_data": {
            "age": 32,
            "gender": "male",
            "gift_name": "꽃",
            "gift_price": 23333,
            "person_name": "김영삼",
            "received_at": "2026-08-19",
            "relationship": "친구"
          }
        },
        "properties": {
          "gift_data": {
            "$ref": "#/components/schemas/GiftData"
          }
        },
        "required": [
          "gift_data"
        ],
        "title": "GiftDataRequest",
        "type": "object"
      },
      "GiftRecommendationInfo": {
        "description": "실제 Qwen 추천 결과와 사용자에게 보낼 감사 메시지.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "message": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "recommend_gift": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimpleGiftRecommendationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "SUCCESS"
          }
        },
        "title": "GiftRecommendationInfo",
        "type": "object"
      },
...
      "ImageRequest": {
        "description": "이미지 분석 API의 요청 본문. 현재는 S3 HTTP(S) 주소를 받습니다.",
        "properties": {
          "image_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Image Url",
            "type": "string"
          }
        },
        "required": [
          "image_url"
        ],
        "title": "ImageRequest",
        "type": "object"
      },
      "PreparedData": {
        "description": "선물 기록·캘린더·알림 준비 함수의 공통 결과.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "SUCCESS"
          }
        },
        "title": "PreparedData",
        "type": "object"
      },
...
      "RecommendRequest": {
        "description": "추천만 단독으로 요청합니다.\n\n나이·가격대·카테고리·성별만으로도 추천이 나옵니다. 사용자가 확인 화면에서\n조건을 바꿔 다시 추천받을 때, 이미지 분석을 다시 돌릴 이유가 없기 때문입니다.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "budget_max": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "budget_min": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Min"
          },
          "categories": {
            "description": "지정하면 이 안에서만 추천합니다",
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Categories",
            "type": "array"
          },
          "dislikes": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Dislikes",
            "type": "array"
          },
          "event": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender",
            "default": "unknown"
          },
          "gift_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "받은 것. 답례 추천일 때",
            "title": "Gift Name"
          },
          "gift_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gift Price"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Interests",
            "type": "array"
          },
          "person_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Name"
          },
          "relationship": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          }
        },
        "title": "RecommendRequest",
        "type": "object"
      },
      "RecommendResponse": {
        "description": "추천 단독 요청의 결과.",
        "properties": {
          "recommend_gift_info": {
            "$ref": "#/components/schemas/GiftRecommendationInfo"
          }
        },
        "required": [
          "recommend_gift_info"
        ],
        "title": "RecommendResponse",
        "type": "object"
      },
...
      "TaskStatus": {
        "description": "각 비동기 준비 작업의 성공/실패 상태.",
        "enum": [
          "SUCCESS",
          "ERROR"
        ],
        "title": "TaskStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "받은 선물과 관계 맥락을 바탕으로 답례 가격대와 카테고리를 추천합니다.",
    "title": "Giftie AI 추천 서비스",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent/confirm": {
      "post": {
        "description": "사용자가 확인 화면에서 검토·수정한 결과를 확정하고 캘린더에 등록합니다.\n\n``/from-image`` 나 ``/from-gift-data`` 는 일정을 등록하지 않고 초안까지만 만듭니다.\n잘못 추출된 일정이 사용자 캘린더에 바로 박히면 되돌리기 어렵기 때문입니다.\n\n이 서비스는 상태를 보관하지 않으므로, 백엔드가 직전 응답을 들고 있다가\n사용자 수정본과 함께 그대로 되돌려주면 됩니다.\n\nArgs:\n    request: 사용자 수정본이 반영된 확정 요청.\n\nReturns:\n    확정된 기록, 캘린더 등록 결과, 알림 예약.",
        "operationId": "confirm_api_v1_agent_confirm_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-KEY",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "confirmDraft": {
                  "summary": "사용자 검토 후 확정",
                  "value": {
                    "approved": true,
                    "gift_data": {
                      "gift_name": "스타벅스 케이크",
                      "gift_price": 35000,
                      "person_name": "김민수",
                      "received_at": "2026-08-19",
                      "relationship": "대학 동기",
                      "target_date": "2026-09-10"
                    },
                    "register_calendar": true,
                    "workflow_id": "from-previous-response-workflow-id"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API 키 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 데이터 검증 실패"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "내부 처리 오류"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bedrock 등 외부 서비스 오류"
          }
        },
        "summary": "Confirm",
        "tags": [
          "gift-agent"
        ]
      }
    },
    "/api/v1/agent/from-gift-data": {
      "post": {
        "description": "구조화된 선물데이터로 네 가지 준비 작업을 실행합니다.\n\nArgs:\n    request: 백엔드가 전달한 ``gift_data`` 요청 본문.\n\nReturns:\n    선물 기록, 캘린더, 알림, 추천/메시지를 합친 응답.",
        "operationId": "prepare_from_gift_data_api_v1_agent_from_gift_data_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-KEY",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "multipleRecords": {
                  "description": "계좌 거래내역처럼 여러 건이면 records를 함께 보냅니다.",
                  "summary": "여러 건의 기록",
                  "value": {
                    "gift_data": {
                      "gift_name": "축의금",
                      "gift_price": 200000,
                      "person_name": "최은비",
                      "received_at": "2026-08-19",
                      "records": [
                        {
                          "confidence": 1.0,
                          "direction": "received",
                          "gift_name": "축의금",
                          "person_name": "김도윤",
                          "price": 100000,
                          "received_at": "2026-08-19",
                          "record_id": "r0",
                          "record_type": "money",
                          "selected": true
                        },
                        {
                          "confidence": 1.0,
                          "direction": "received",
                          "gift_name": "축의금",
                          "person_name": "박서준",
                          "price": 50000,
                          "received_at": "2026-08-19",
                          "record_id": "r1",
                          "record_type": "money",
                          "selected": true
                        }
                      ]
                    }
                  }
                },
                "singleGift": {
                  "description": "일반적인 직접 입력입니다. records를 보내지 않습니다.",
                  "summary": "단건 선물데이터",
                  "value": {
                    "gift_data": {
                      "age": 29,
                      "gender": "female",
                      "gift_name": "스타벅스 케이크",
                      "gift_price": 35000,
                      "person_name": "김민수",
                      "received_at": "2026-08-19",
                      "relationship": "대학 동기",
                      "target_date": "2026-09-10"
                    }
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/GiftDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiftAgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API 키 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 데이터 검증 실패"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "내부 처리 오류"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bedrock 등 외부 서비스 오류"
          }
        },
        "summary": "Prepare From Gift Data",
        "tags": [
          "gift-agent"
        ]
      }
    },
    "/api/v1/agent/from-image": {
      "post": {
        "description": "S3 이미지 주소를 선물데이터로 변환한 뒤 네 작업을 실행합니다.\n\n``MODEL_BACKEND=bedrock`` 또는 ``vllm``에서는 실제 이미지 분석을 수행하고,\n``mock``에서는 외부 모델 없이 고정 결과로 전체 연동 흐름을 검증합니다.",
        "operationId": "prepare_from_image_api_v1_agent_from_image_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-KEY",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "s3PresignedUrl": {
                  "summary": "S3 presigned 이미지 URL",
                  "value": {
                    "image_url": "https://example-bucket.s3.ap-northeast-2.amazonaws.com/u1/gift.png?X-Amz-Signature=example"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiftAgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API 키 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 데이터 검증 실패"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "내부 처리 오류"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bedrock 등 외부 서비스 오류"
          }
        },
        "summary": "Prepare From Image",
        "tags": [
          "gift-agent"
        ]
      }
    },
    "/api/v1/agent/recommend": {
      "post": {
        "description": "추천만 단독으로 실행합니다.\n\n나이·가격대·카테고리·성별만으로도 추천이 나옵니다. 사용자가 확인 화면에서\n조건을 바꿔 다시 추천받을 때, 이미지 분석과 캘린더·알림을 다시 돌릴 이유가 없습니다.\n\nArgs:\n    request: 추천 조건.\n\nReturns:\n    추천 가격대·카테고리·실제 상품·메시지와 그 근거.",
        "operationId": "recommend_api_v1_agent_recommend_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-KEY",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "recommendOnly": {
                  "summary": "추천 조건만 다시 실행",
                  "value": {
                    "age": 32,
                    "budget_max": 27000,
                    "budget_min": 18000,
                    "categories": [
                      "꽃·식물"
                    ],
                    "gender": "male",
                    "gift_name": "꽃",
                    "gift_price": 23333,
                    "person_name": "김영삼",
                    "relationship": "친구"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API 키 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 데이터 검증 실패"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "내부 처리 오류"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bedrock 등 외부 서비스 오류"
          }
        },
        "summary": "Recommend",
        "tags": [
          "gift-agent"
        ]
      }
    }
  }
}